requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# RKLLM Runtime (will use system library via ctypes)
# numpy for numerical operations (1.26+ required for Python 3.12)
numpy>=1.26.0

# Additional utilities
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.32.3  # For testing
huggingface_hub>=0.20.0  # For HF model discovery

# Development and debugging
python-json-logger==2.0.7

# Stable Diffusion Dependencies
diffusers>=0.24.0
transformers>=4.36.0
scipy>=1.11.0
accelerate>=0.25.0
diffusers
transformers
pillow
# rknn-toolkit-lite2  # Install manually or via board-specific wheel
